Crud userbook works

+++ src/app/api/user-books/route.ts
@@ -1,54 +1,21 @@
 import { Logger } from "@/src/logger";
 import { UserBook, UserBookData } from "@/src/models";
-import {
-  BookDTO,
-  GenreDTO,
-  GoodreadsDataDTO,
-  ReadingStatusDTO,
-} from "@/src/models/dto";
 import { IResponse } from "@/src/models/dto/response";
-import UserBookDTO, {
+import {
   CreateUserBookBody,
-  GetUserBooksResponseDTO,
+  DeleteUserBookBody,
   UpdateUserBookBody,
 } from "@/src/models/dto/userBookDTO";
 import { GetAxiosInstance } from "@/src/utils/apiUtils";
 import { NextRequest, NextResponse } from "next/server";
-
-const userBookDataDTOFromResponse = (
-  userBookDataDTO: GetUserBooksResponseDTO
-) =>
-  new UserBookData(
-    {
-      book: BookDTO.FromResponse(userBookDataDTO.book_data.book),
-      mainGenre: GenreDTO.FromResponse(userBookDataDTO.book_data.main_genre),
-      subgenres: userBookDataDTO.book_data.subgenres?.map((subgenre) =>
-        GenreDTO.FromResponse(subgenre)
-      ),
-    },
-    ReadingStatusDTO.FromResponse(userBookDataDTO.reading_status),
-    userBookDataDTO.user_book_id,
-    userBookDataDTO.user_id,
-    userBookDataDTO.suggestion_source,
-    userBookDataDTO.user_comments,
-    userBookDataDTO.date_added,
-    userBookDataDTO.user_rating,
-    userBookDataDTO.reading_start_date,
-    userBookDataDTO.reading_finish_date,
-    userBookDataDTO.is_deleted,
-    userBookDataDTO.is_favorite,
-    GoodreadsDataDTO.FromResponse(userBookDataDTO.book_data.goodreads_data)
-  );
 
 export async function GET(
   req: NextRequest
 ): Promise<NextResponse<IResponse<UserBookData[]>>> {
   try {
     const axios = GetAxiosInstance(req);
-    const response = await axios.get<GetUserBooksResponseDTO[]>("/user-book");
-    const userBookData = response.data?.map((getUserBooksResponseDTO) =>
-      userBookDataDTOFromResponse(getUserBooksResponseDTO)
-    );
+    const response = await axios.get<UserBookData[]>("/user-book");
+    const userBookData = response.data;
     const result: IResponse<UserBookData[]> = {
       result: userBookData,
     };
@@ -66,13 +33,12 @@ export async function POST(req: NextRequest) {
   try {
     createUserBookBody = await req.json();
     const axios = GetAxiosInstance(req);
-    const response = await axios.post<UserBookDTO>(
+    const response = await axios.post<UserBook>(
       "/user-book",
       createUserBookBody
     );
 
-    const userBookDTO = response.data;
-    const userBook = UserBookDTO.FromResponse(userBookDTO);
+    const userBook = response.data;
 
     return NextResponse.json(
       {
@@ -99,11 +65,11 @@ export async function PATCH(
   try {
     updateUserBookBody = await req.json();
     const axios = GetAxiosInstance(req);
-    const response = await axios.patch<UserBookDTO>(
+    const response = await axios.patch<UserBook>(
       "/user-book",
       updateUserBookBody
     );
-    const userBook: UserBook = UserBookDTO.FromResponse(response.data);
+    const userBook: UserBook = response.data;
     return NextResponse.json({ result: userBook }, { status: 200 });
   } catch (error: any) {
     Logger.error("Error updating user book", {
@@ -123,11 +89,13 @@ export async function DELETE(
   try {
     const body = await req.json();
     userBookId = body.userBookId;
+    if (!userBookId) throw new Error("userBookId is null");
     const axios = GetAxiosInstance(req);
+    const data: DeleteUserBookBody = {
+      userBookId: userBookId,
+    };
     const response = await axios.delete<IResponse<void>>("/user-book", {
-      data: {
-        user_book_id: userBookId,
-      },
+      data,
     });
     return NextResponse.json(response.data, { status: 200 });
   } catch (error: any) {
